处理当点击 同步数据按钮时 有些数据为空出现的问题

+++ app/models/activity_sign_up.rb
@@ -2,11 +2,13 @@ class ActivitySignUp < ActiveRecord::Base
 
   def self.update_user_activity_sign_ups(params)
     #ActivitySign_up.transaction do
-    ActivitySignUp.delete_all(:user_name => params[:user_name])
-    activity_sign_ups = params[:activity_sign_ups]
-    activity_sign_ups.each do |activity_sign_up|
-      activity_sign_up_info = ActivitySignUp.new(activity_sign_up)
-      activity_sign_up_info.save
+    if params[:activity_sign_up]
+      ActivitySignUp.delete_all(:user_name => params[:user_name])
+      activity_sign_ups = params[:activity_sign_ups]
+      activity_sign_ups.each do |activity_sign_up|
+        activity_sign_up_info = ActivitySignUp.new(activity_sign_up)
+        activity_sign_up_info.save
+      end
     end
     #end
   end
