+++ app/models/bid_sign_up.rb
@@ -2,11 +2,13 @@ class BidSignUp < ActiveRecord::Base
 
   def self.update_user_bid_sign_ups(params)
     #BidSignUp.transaction do
-    BidSignUp.delete_all(:user_name => params[:user_name])
-    bid_sign_ups = params[:bid_sign_ups]
-    bid_sign_ups.each do |bid_sign_up|
-      bid_sign_up_info = BidSignUp.new(bid_sign_up)
-      bid_sign_up_info.save
+    if params[:BidSignUp]
+      BidSignUp.delete_all(:user_name => params[:user_name])
+      bid_sign_ups = params[:bid_sign_ups]
+      bid_sign_ups.each do |bid_sign_up|
+        bid_sign_up_info = BidSignUp.new(bid_sign_up)
+        bid_sign_up_info.save
+      end
     end
     #end
   end
@@ -18,14 +20,14 @@ class BidSignUp < ActiveRecord::Base
   end
 
   def self.update_winner_info(params)
-      params_winner_infos = params[:winner_info]
-      params_winner_info = params_winner_infos.first
-      winner_info = BidSignUp.where(:user_name => params_winner_info[:user_name],
-                                    :activity_name => params_winner_info[:activity_name],
-                                    :bid_name => params_winner_info[:bid_name],
-                                    :phone => params_winner_info[:phone]).first
-      winner_info.is_winner = true
-      winner_info.save
+    params_winner_infos = params[:winner_info]
+    params_winner_info = params_winner_infos.first
+    winner_info = BidSignUp.where(:user_name => params_winner_info[:user_name],
+                                  :activity_name => params_winner_info[:activity_name],
+                                  :bid_name => params_winner_info[:bid_name],
+                                  :phone => params_winner_info[:phone]).first
+    winner_info.is_winner = true
+    winner_info.save
   end
 
 end
